error

+++ Control/CError.php
@@ -2,6 +2,12 @@
 
 class CError
 {
+    private static function authorizer():bool{
+
+        if(FSession::isLogged() && (FSession::getUserLogged()->getType()!='Administrator'))throw new Exception("you don't have authorization");
+        return CManageUser::callLogin();
+    }
+
     /**
      * @param Exception $error
      * @param string|NULL $message
@@ -14,12 +20,38 @@ class CError
     }
 
     public static function read(){
-        FDbH::readErrors();
-        //richiama la view
+        try{
+            if(self::authorizer()){
+                if(FSession::isLogged()){
+                    $user=FSession::getUserLogged();
+                    $profileImg=FDbH::loadMultiFile($user,MappingPathFile::nameUserMain(),MappingPathFile::dirUserDefault(),MappingPathFile::nameUserMain(),0.2);
+                }
+                else{
+                    $user=null;
+                    $profileImg=null;
+                }
+                $error=FDbH::readErrors();
+                $view=new VErrorSystem();
+                $view->show($error, $user, $profileImg);
+            }
+            else throw new Exception("you don't have authorization");
+        }
+        catch (Exception $e){
+            CError::store($e,"ci scusiamo per il disaggio !!! non è stato possibile visualizzare la pagina degli errori , verificare di possedere le autorizzazioni necessarie");
+        }
     }
 
     public static function delete(){
-        FDbH::deleteErrors();
+        try{
+            if(self::authorizer()){
+                FDbH::deleteErrors();
+                header('Location: /Livent/Error/');
+            }
+            else throw new Exception("you don't have authorization");
+        }
+        catch (Exception $e){
+            CError::store($e,"ci scusiamo per il disaggio !!! non è stato possibile visualizzare la pagina degli errori , verificare di possedere le autorizzazioni necessarie");
+        }
     }
 
     public static function getFile(){
